Replace usage of `phdr.p_paddr` with `phdr.p_vaddr`
According to ELF specification:
```
On systems for which physical addressing is relevant, this member
is reserved for the segment's physical address. Because System V
ignores physical addressing for application programs, this member
has unspecified contents for executable files and shared objects.
```
Furthermore:
```
On systems for which physical addressing is relevant, this member is
reserved for the segment's physical address. Because System V ignores
physical addressing for application programs, this member has
unspecified contents for executable files and shared objects.
```

Thus, even though in practice `p_paddr` is almost always equal to
`p_vaddr`, ensure consistency with the specification and use `p_vaddr`
instead.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #64